Moved code relative to openid4vp in the relative package
